+++ src/node/plugins/index.ts
@@ -1,4 +1,4 @@
-import { Plugin, defineConfig, mergeConfig } from "vite";
+import { defineConfig, mergeConfig, PluginOption } from "vite";
 import fsx from "fs-extra";
 import path from "path";
 import react from "@vitejs/plugin-react";
@@ -12,12 +12,13 @@ import { parseApi } from "../utils/api.js";
 import { compileSync } from "@mdx-js/mdx";
 const pkg = fsx.readJSONSync(PKG_JSON_PATH);
 
-export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
-  const provider = await virtualProvider(config);
+export const docit = async (
+  config: ResolvedUserConfig
+): Promise<PluginOption> => {
   const instance = Core.getInstance(config);
-  const { sidebars, routes, appData, sandboxes } = await instance.prepare();
+  const virtualPlugins = await instance.prepare();
 
-  const plugin: Plugin = {
+  const docitPlugin: PluginOption = {
     name: "vite-plugin-docit",
     enforce: "pre",
     config: () => {
@@ -25,17 +26,17 @@ export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
         root: config.base,
         base: config.publicPath,
         optimizeDeps: {
-          include: [
-            "react",
-            "react-dom",
-            "styled-components",
-            "shallowequal",
-            "hoist-non-react-statics",
-            "react-is",
-            "lodash-es",
-            "highlight.js",
-            "qrcode",
-          ],
+          // include: [
+          //   "react",
+          //   "react-dom",
+          //   "styled-components",
+          //   "shallowequal",
+          //   "hoist-non-react-statics",
+          //   "react-is",
+          //   "lodash-es",
+          //   "highlight.js",
+          //   "qrcode",
+          // ],
           exclude: ["@mdx-js/react"],
         },
         build: {
@@ -72,23 +73,24 @@ export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
       }
     },
   };
+  const mdxPlugin = await mdx(config);
+  const nodeResolvePlugin = nodeResolve({
+    moduleDirectories: [
+      "node_modules",
+      `./node_modules/${pkg.name}/node_modules`,
+    ],
+  }) as PluginOption;
+  const providerPlugin = virtualProvider(config);
+  const reactPlugin = react({
+    jsxRuntime: "classic",
+  });
 
   return [
-    plugin,
-    appData,
-    routes,
-    sidebars,
-    provider,
-    sandboxes,
-    ...(await mdx(config)),
-    ...(react({
-      jsxRuntime: "classic",
-    }) as Plugin[]),
-    nodeResolve({
-      moduleDirectories: [
-        "node_modules",
-        `./node_modules/${pkg.name}/node_modules`,
-      ],
-    }),
+    docitPlugin,
+    virtualPlugins,
+    providerPlugin,
+    reactPlugin,
+    mdxPlugin,
+    nodeResolvePlugin,
   ];
 };