+++ src/node/plugins/virtual/provider.ts
@@ -1,15 +1,15 @@
-import { Plugin } from 'vite';
-import { ResolvedUserConfig } from '../../types.js';
+import { PluginOption } from "vite";
+import { ResolvedUserConfig } from "../../types.js";
 
-const virtualProvider = async (config: ResolvedUserConfig): Promise<Plugin> => {
-  return Promise.resolve({
-    name: 'virtual:vite-plugin-virtual/virtual:provider',
+const virtualProvider = (config: ResolvedUserConfig): PluginOption => {
+  return {
+    name: "virtual:vite-plugin-virtual/virtual:provider",
     resolveId(id) {
-      if (id === 'virtual:provider') {
+      if (id === "virtual:provider") {
         return config.providerPath;
       }
     },
-  });
+  };
 };
 
 export { virtualProvider };