+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - lock some deps from dependencies list
+- change mdast related lib to remark ecosystem
 
 ## [0.10.2] - 2022-08-02